now the web app depends on zkee.
in spi form, popups were replaced by modal windows due to some client
problem with zkee (TypeError on invoke in the browser).

in customers tree, when an item is disabled, now the corresponding tree
cell background is colored in red instead of being detached from the
tree. that, to improve the performance when manipulating thousands of
customers.